backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-dotenv
pydantic
langchain
langchain-google-genai
langchain-core
google-generativeai
